MySQL Database Attachment
Hooking up particle page to SQL database hosted on webserver. Code will generate particles, pick a random particle, send its x and y coords to the database to be pulled by a client side python script. These coordinates will be used to set the initial width and height for a game of life configuration.

This will eventually be used in a generative art project.